Tests for single valued criterion.
